+++ custom_components/maestro/http/request.py
@@ -1,0 +1,7 @@
+def RequestBuilder(self, ConfigurationId: str, Commands: str):
+    return  {
+                "ModelId": self.ModelId,
+                "ConfigurationId":ConfigurationId,
+                "SensorSetTypeId":self.SensorSetTypeId,
+                "Commands":Commands
+            }
