+++ custom_components/maestro/types/mode.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class ModeEnum(Enum):
+    manual = 0,
+    auto = 1,
+    overnight = 2,
+    comfort = 3,
+    turbo = 4
