+++ custom_components/manifest.json
