+++ hacs.json
@@ -1,4 +1,4 @@
 {
   "name": "Maestro MCZ (cloud)",
-  "content_in_root": true
+  "content_in_root": false
 }
