--- maestro/http/request.py
@@ -1,7 +1,0 @@
-def RequestBuilder(self, ConfigurationId: str, Commands: str):
-    return  {
-                "ModelId": self.ModelId,
-                "ConfigurationId":ConfigurationId,
-                "SensorSetTypeId":self.SensorSetTypeId,
-                "Commands":Commands
-            }
